[src/verifier] Change C header to CXX header
Filenames should be the class name they define.
Keep report.h for compatibility
