+++ src/verifier/Report.cpp
@@ -2,7 +2,7 @@
 // Copyright (c) 2018, The Regents of the University of California (Regents).
 // All Rights Reserved. See LICENSE for license details.
 //------------------------------------------------------------------------------
-#include <report.h>
+#include <Report.hpp>
 #include <cstring>
 #include <iomanip>
 #include <iostream>
